backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
python-multipart==0.0.6
structlog==23.2.0
httpx==0.25.2
python-socketio==5.9.0
websockets==12.0
docker==7.0.0
requests-unixsocket2==0.4.2

# AI Agent 依赖 (简化版本)
langchain>=0.1.0
langgraph>=0.0.40
openai>=1.0.0
chromadb>=0.4.18

# 搜索服务依赖
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
